+++ swagmodels/threshold_crossing_by_m_o_top_n_config.go
@@ -23,6 +23,9 @@ type ThresholdCrossingByMOTopNConfig struct {
 	// Enum: [min max avg]
 	Aggregator string `json:"aggregator,omitempty"`
 
+	// A value of true will have the aggregation request execute on all data regardless of whether it has been cleaned or not
+	IgnoreCleaning bool `json:"ignoreCleaning,omitempty"`
+
 	// Time boundary for the metrics under consideration using the ISO-8601 standard
 	// Required: true
 	Interval *string `json:"interval"`